Update Linux FAQ section
Maybe a repeat of #20 but the Linux tips in the FAQ section of the guide are a bit out of date with the state of things on the Linux desktop, and in the absence of official resources, community members have filled in the gaps. 

I maintain the unofficial SteamOS 3 guide and administer the unofficial Linux support in your Discord, so it stands to reason that there should be some mention of that. The thing I spend most of my time doing in that channel is reposting the URL to the guide and so putting it in the FAQ _hopefully_ will reduce some of that.